+++ src/main/java/com/aipiabackend/order/service/dto/OrderLineCreateCommand.java
@@ -1,11 +1,9 @@
 package com.aipiabackend.order.service.dto;
 
-import java.math.BigDecimal;
-
 public record OrderLineCreateCommand(
     Long productId,
     Integer productQuantity,
-    BigDecimal productPrice,
-    BigDecimal amount
+    Long productPrice,
+    Long amount
 ) {
 }
